Backend/requirements.txt
Flask == 1.1.2
Flask_RESTful_TheodoFork == 0.3.7
cryptography == 3.3.1
mysql_connector == 2.2.9
